backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
openai==1.51.0
groq==0.11.0
# Pin httpx to 0.27.x due to groq client compatibility with 'proxies' kwarg
httpx==0.27.2
wikipedia==1.4.0
google-search-results==2.4.2
reportlab==4.2.2
python-docx==1.1.2
requests==2.32.3
tqdm==4.66.5
gradio-client==0.15.0
duckduckgo-search==3.9.13
PyMuPDF==1.26.7
beautifulsoup4==4.14.3
